requirements.txt
streamlit
pandas
numpy
scikit-learn
geopandas
matplotlib
seaborn
joblib
fpdf2
google-generativeai
python-dotenv